Keep existing attachments by fetching message if required
Descriptions are sent in the `attachments` list, which is also used for removing existing attachments on edit.
Many types (like partial messages, webhooks) do not carry information about previous attachments,
which would resulted in them being removed if new attachments are added on edit.

To fix this, the existing message is fetched beforehand and its attachments are used as the baseline,
if new files are also being added.

While this does introduce a new API request in some cases, it is required for using attachments descriptions (as all
attachments have to be provided in the request), and will be required for API v10 at the latest regardless.