Remove unused code
Explanation:
If we got to the now removed `member = ...` part, we know that at least two conditions
would've had to hold: `guild is not None` and `member_data`. If that's the case though,
`raw.member` would've already been set beforehand, in which case we couldn't even reach
the removed code.