Don't copy kwargs in __new__
While the existing code works fine, I'd rather not create and store a copy of the kwargs
on every instantiation, since it's really only required for creating a copy of the
decorator-created item in the class and assigning that copy to the new instance.

Manually created items don't need to be copied at any point, so storing the
kwargs used to create them would be pointless.