Fixed a rare command sync bug (#254)
* Fixed a rare command sync bug

It was impossible to register commands of different types with the same name. The reason was one of the internal functions that compares old and new commands. It mapped them like {name: cmd} instead of {(name, type): cmd}

* Remove "Type migration" because it's no longer an issue

Imagine we have command A of one type and command B of another type but with the same name.
Previously, API didn't allow to overwrite A with B without forcing users to make an extra step of deleting A and inserting B afterwards.

* Fixed formatting