feat: configurable max wait time for ratelimits
Currently when doing a http call there is no way to know when exactly it
will be executed due to rate limits. While in some cases a delay of a
few seconds is not a big deal, other limits (most notably are guild wide
limits of up to 24h hours on roles and emoji) are so long that it is not
reasonable to wait for them. As this creates a terrible end user
experience (bot is not doing anything, did it break?)

This adds a `max_ratelimit_wait` kwarg to the client creation which will
raise a RatelimitTooLong exception on any endpoint if the time to wait
would be too long.