docs: clarify `default_member_permissions` override
Additional note regarding the current client behavior:

When a command specifies `default_member_permissions`, only users with all those permissions
may invoke the command.
If a moderator now modifies the *command-specific* permissions in some way, making them un-synced,
`default_member_permissions` will be ignored *entirely* until the command is synced again.

As an example, consider users A, B, C, with A having ban perms and B + C with no special perms,
as well as a command that requires ban perms using `default_member_permissions`.
By default, A can invoke the command, and B + C can't.
Currently, `default_member_permissions` is not reflected in the settings UI,
which means that everything is shown as 'allowed' by default.
If a moderator now visits the guild settings and allows C to use this command,
un-syncing the command permissions, B will now also be able to use the command
if the moderator didn't adjust the permissions further.

This should hopefully be fixed once the field is reflected in the settings UI (as originally proposed).