fix!: change ignore_timeout default to False
after seeing quite a few usecases of this method in the wild we've
decided to change the default of `ignore_timeout` to False.

While this is a breaking change, it matches how the client and API both
calculate permissions when a user is timed out and we feel we should
mirror that behavior.