feat: use pdm for dependency management (#836)
This migrates the final portion of the build configuration to PEP 621,
but continues to keep setuptools around as the build backend, for now.

This change brings about an easier way to manage dependencies, and no
more reliance on having multiple requirement files. This also means that 
we don't have to bundle multiple requirements files in our builds.

Additionally, this should not change our wheel or sdist metadata in any
noticeable way. The largest impact this has by far is simply cleaning up
our dependencies across multiple files and introduces a manager to help
manage our dependencies.

Signed-off-by: arl <me@arielle.codes>
Co-authored-by: shiftinv <me@shiftinv.cc>