fix: copy `ParamInfo` instances from signatures before processing them (#946)
Currently it's not recommended to reference the same `ParamInfo`
instance in multiple signatures, as it results in weird behavior of
`LocalisationStore`. However, this is not documented and it's not
obvious for an average user. Sometimes creating a single `ParamInfo`
instance for multiple slash commands is very tempting, e.g. a choice
between `On` and `Off` with localized labels. This PR allows to safely
do that.