fix(docs): use environment variable instead of config context for `html_baseurl` (#1049)
Fixes readthedocs builds, which broke some time within the last 24 hours
due to an upstream change which resulted in the `html_baseurl`
variable being set to `""`. 

This has the (positive) side-effect of `<link rel="canonical" />` no
longer always pointing to `stable`, and instead using the version that's
being built. The same applies to the opensearch config.