docs: rewrite CONTRIBUTING.md (#1098)
A general overhaul of the entire contribution guide, hopefully making it
easier to follow and understand. Also includes the documentation side of
#196, i.e. outlining the commit/PR naming convention.

This removes the detailed example which, while insightful, was way too
lengthy for new contributors. We could maybe re-add it on a separate
page in the future, but for now it's too much for this guide.