perf: prefer `parent.channel` over `get_channel()`
I don't think there are even cases where `resolved["messages"]` can contain a message
that *isn't* in the current channel, so `channel_id == parent.channel.id` should always match.
Still, we fall back to `guild.get_channel()` in case it doesn't.

Additionally, we also fall back in the same way to `state.get_channel()` to make use of the DM cache if `parent.channel` is
a `PartialMessageable`, since that means it's a dm/group channel, for which support will be added in another PR.