fix: handle refreshing views with premium buttons correctly
This updates the workaround added in 41618b1cd86ea5904fb4d32eb0321eb462375720,
which attempts to preserve the row of url buttons. This already technically worked
for premium buttons too, but only due to a bug in that previous implementation.

Still, this remains an awful workaround. It may be easier to only update the components
directly, and leave `self.children` as is, ignoring additions/removals (which shouldn't happen here
in the first place). Alternatively, the new `id` field could also work, though it's currently still
part of that other thing and not public.