fix: provide a fallback version when the version cannot be determined
Without this setting, installation of disnake from a source distribution
without source control metadata will fail, because there will be no way
to determine the version of disnake without that metadata.

Using a fallback ensures that in such an installation method (such as
from a GitHub source archive), disnake can still be installed, though
the version metadata may be inaccurate.

Compared to not allowing installation at all, this is a better user
experience, despite the incorrect version information.