fix(AppInfo): allow nullable integration_types_config to be able to clear config
To remove eg user installations, that dictionary mustn't be provided
so if we want to clear it we must provide the absense of that field
This supports None as a way to remove support for a specific tag

Much as it may be our place to provide TypeErrors for users, this endpoint
is filled with some legacy cruft from Discord such as install_params and
custom_install_url. These two fields *are* mutually exclusive, but in
some cases, the user needs to provide a dictionary to one key, and another
dictionary to another key...

It is easier and overall safer for us to provide the majority of this method
directly to the user, and not doing any checking for handling.