fix: base custom search scorer on base title score
this is by no means perfect; long story short, sphinx changed
the search scorer algorithm to actually take the configured `title`
score (15) into account for matching titles, instead of a fixed 100.
This, in turn, meant that any and all page matches would now be ranked at the bottom,
since the custom search scorer inflates the score of all `disnake.*` objects to
a score right below 100.

To resolve this, the scorer now assigns these API results a score of
14-15, i.e. right below pages with matching titles again.
API results with a longer or later submatch continue to receive a higher score penalty,
and are therefore closer to 14 rather than 15.